Fix issue with comparison of string constants
String constants were compare literally, as part of constant propagation.
This means that for example `"\a" == "a"` was compiled to `false` even though the two represent the same strings.
Unicode and hex escape codes are other examples where literal comparison is not correct.

Implemented a 3-way string comparison: equality returns Some(true), Some(false), or None where the result is not determined.